+++ gson/src/main/java/feign/gson/GsonCodec.java
@@ -1,7 +1,6 @@
 package feign.gson;
 
 import com.google.gson.Gson;
-import com.google.gson.JsonIOException;
 import feign.RequestTemplate;
 import feign.Response;
 import feign.codec.Decoder;
@@ -9,40 +8,30 @@ import feign.codec.Encoder;
 
 import javax.inject.Inject;
 import java.io.IOException;
-import java.io.Reader;
 import java.lang.reflect.Type;
 
-import static feign.Util.ensureClosed;
-
+/**
+ * @deprecated use {@link GsonEncoder} and {@link GsonDecoder} instead
+ */
+@Deprecated
 public class GsonCodec implements Encoder, Decoder {
-  private final Gson gson;
+  private final GsonEncoder encoder;
+  private final GsonDecoder decoder;
 
   public GsonCodec() {
     this(new Gson());
   }
 
   @Inject public GsonCodec(Gson gson) {
-    this.gson = gson;
+    this.encoder = new GsonEncoder(gson);
+    this.decoder = new GsonDecoder(gson);
   }
 
   @Override public void encode(Object object, RequestTemplate template) {
-    template.body(gson.toJson(object));
+    encoder.encode(object, template);
   }
 
   @Override public Object decode(Response response, Type type) throws IOException {
-    if (response.body() == null) {
-      return null;
-    }
-    Reader reader = response.body().asReader();
-    try {
-      return gson.fromJson(reader, type);
-    } catch (JsonIOException e) {
-      if (e.getCause() != null && e.getCause() instanceof IOException) {
-        throw IOException.class.cast(e.getCause());
-      }
-      throw e;
-    } finally {
-      ensureClosed(reader);
-    }
+    return decoder.decode(response, type);
   }
 }